requirements.txt
Flask
Flask-Cors
Flask-SocketIO
eventlet
gunicorn
paho-mqtt
